+++ .eslintignore
@@ -6,8 +6,6 @@ node_modules
 .env
 .env.*
 !.env.example
-
-# Ignore files for PNPM, NPM and YARN
 pnpm-lock.yaml
 package-lock.json
 yarn.lock
