+++ .gitignore
@@ -1,2 +1,11 @@
-build
+.DS_Store
 node_modules
+/build
+/.svelte-kit
+/package
+.env
+.env.*
+!.env.example
+.output
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
