+++ package.json
@@ -6,42 +6,41 @@
   "version": "1.0.0",
   "license": "MIT",
   "author": "Sean Lingren <sean@lingrino.com> (https://lingrino.com/)",
+  "type": "module",
   "scripts": {
-    "dev": "snowpack dev",
-    "build": "snowpack build",
+    "dev": "vite dev",
+    "build": "vite build",
+    "preview": "vite preview",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
     "test": "exit 0",
     "lint": "eslint --fix .",
     "lint:check": "eslint .",
     "format": "prettier --write .",
     "format:check": "prettier --check ."
   },
-  "dependencies": {
-    "@tailwindcss/forms": "^0.5.3",
-    "svelte": "^3.58.0",
-    "tailwindcss": "^3.3.1"
-  },
   "devDependencies": {
-    "@snowpack/plugin-build-script": "^2.1.0",
-    "@snowpack/plugin-dotenv": "^2.2.0",
-    "@snowpack/plugin-svelte": "^3.7.0",
-    "@snowpack/plugin-typescript": "^1.2.1",
-    "@tsconfig/svelte": "^3.0.0",
-    "@types/snowpack-env": "^2.3.4",
+    "@sveltejs/adapter-cloudflare": "^2.2.0",
+    "@sveltejs/kit": "^1.5.0",
+    "@sveltejs/vite-plugin-svelte": "^2.0.4",
+    "@tailwindcss/forms": "^0.5.3",
+    "@tsconfig/svelte": "^4.0.1",
     "@typescript-eslint/eslint-plugin": "^5.57.1",
     "@typescript-eslint/parser": "^5.57.1",
     "autoprefixer": "^10.4.14",
     "eslint": "^8.37.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-svelte3": "^4.0.0",
     "postcss": "^8.4.21",
-    "postcss-import": "^15.1.0",
-    "postcss-load-config": "^4.0.1",
     "prettier": "^2.8.7",
     "prettier-plugin-svelte": "^2.10.0",
-    "snowpack": "^3.8.8",
+    "svelte": "^3.58.0",
+    "svelte-check": "^3.0.1",
     "svelte-preprocess": "^5.0.3",
-    "typescript": "^4.9.5"
+    "tailwindcss": "^3.3.1",
+    "tslib": "^2.4.1",
+    "typescript": "^5.0.0",
+    "vite": "^4.2.0"
   }
 }
