+++ postcss.config.js
@@ -1,7 +1,6 @@
-const postcssimport = require("postcss-import");
-const tailwindcss = require("tailwindcss");
-const autoprefixer = require("autoprefixer");
-
-module.exports = {
-  plugins: [postcssimport, tailwindcss, autoprefixer],
+export default {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
 };
