--- src/Styles.svelte
@@ -1,5 +1,0 @@
-<style global>
-  @tailwind base;
-  @tailwind components;
-  @tailwind utilities;
-</style>
