--- src/components/Footer.svelte
@@ -1,15 +1,0 @@
-<footer>
-  <div class="py-4 text-center text-lg">
-    Made by <a
-      href="https://lingrino.com"
-      target="_blank"
-      rel="noopener"
-      class="text-blue-600">Sean Lingren</a
-    >&nbsp;&nbsp;•&nbsp;&nbsp;<a
-      href="https://github.com/lingrino/uptime"
-      target="_blank"
-      rel="noopener"
-      class="text-blue-600">Source Code</a
-    >
-  </div>
-</footer>
