--- src/components/Title.svelte
@@ -1,5 +1,0 @@
-<header>
-  <div class="text-center">
-    <p class="text-6xl sm:text-9xl text-green-900">Uptime Calculator</p>
-  </div>
-</header>
