--- src/index.ts
@@ -1,14 +1,0 @@
-import App from "./App.svelte";
-
-const app = new App({
-  target: document.body,
-});
-
-export default app;
-
-if (import.meta.hot) {
-  import.meta.hot.accept();
-  import.meta.hot.dispose(() => {
-    app.$destroy();
-  });
-}
