+++ src/routes/eight-nines/+page.ts
@@ -1,0 +1,5 @@
+import type { PageLoad } from "./$types.js";
+
+export const load = (() => ({
+  uptime: 99.999999,
+})) satisfies PageLoad;
