+++ src/routes/nine-fives/+page.ts
@@ -1,0 +1,5 @@
+import type { PageLoad } from "./$types.js";
+
+export const load = (() => ({
+  uptime: 55.5555555,
+})) satisfies PageLoad;
